moved methods from room_booking back to reservation_hub
