+++ spec/room_booking_spec.rb
@@ -1,0 +1,2 @@
+require_relative 'spec_helper'
+require 'pry'
